requirements.txt
discord.py
flake8